refactor: group related functionality into modules
