+++ packages/core/src/messagingProtocols/proteus/Utility/PrekeyHandler/index.ts
@@ -17,4 +17,4 @@
  *
  */
 
-export * from './SessionIdBuilder';
+export * from './PrekeyHandler';
