+++ packages/core/src/messagingProtocols/proteus/Utility/extractEncryptedAndMissingFromBatchedPayload/index.ts
